move around serialization code
