+++ lib/delayed/backend/active_record.rb
@@ -1,19 +1,5 @@
 require 'active_record'
 
-class ActiveRecord::Base
-  yaml_as "tag:ruby.yaml.org,2002:ActiveRecord"
-
-  def self.yaml_new(klass, tag, val)
-    klass.find(val['attributes']['id'])
-  rescue ActiveRecord::RecordNotFound
-    nil
-  end
-
-  def to_yaml_properties
-    ['@attributes']
-  end
-end
-
 module Delayed
   module Backend
     module ActiveRecord
