+++ lib/delayed/deserialization_error.rb
@@ -1,0 +1,4 @@
+module Delayed
+  class DeserializationError < StandardError
+  end
+end
