+++ lib/delayed/serialization/active_record.rb
@@ -1,0 +1,13 @@
+class ActiveRecord::Base
+  yaml_as "tag:ruby.yaml.org,2002:ActiveRecord"
+
+  def self.yaml_new(klass, tag, val)
+    klass.find(val['attributes']['id'])
+  rescue ActiveRecord::RecordNotFound
+    nil
+  end
+
+  def to_yaml_properties
+    ['@attributes']
+  end
+end
