+++ lib/delayed/worker.rb
@@ -30,6 +30,7 @@ module Delayed
 
     def self.backend=(backend)
       if backend.is_a? Symbol
+        require "delayed/serialization/#{backend}"
         require "delayed/backend/#{backend}"
         backend = "Delayed::Backend::#{backend.to_s.classify}::Job".constantize
       end
